Several unrelated changes in this commit. Fixed emma.build.xml so that it properly excludes tests and my tinker packages from coverage.
Implemented some code that might make qa.build.xml's display functionality work on Linux and Windows, but haven't tested on anything but Mac OS X. It's also kinda gross (lots of duplicated Ant code) but can't figure out how to refactor it without writing my own Ant task, which is total overkill. Also, seems like the display function should really be in each QA tool's build file, but for now I wish to keep the build files as similar to Hackystat's as possible.

Added tests for the stub User and Users resources. Turned out to be harder than I expected, since it required WattDepotClient to start doing authenticated HTTP stuff. Lots of refactoring, so now the tests and even the client code is short and sweet.

Had strange problem with Junit tests for Users resource, saying that the port was already in use when starting up a server for testing. Fixed by not recreating the Server in the static field if it is already there, but not sure that is the correct solution. Needs more investigation.

git-svn-id: https://wattdepot.googlecode.com/svn/trunk@57 962b3a1e-5fa1-11de-82f7-67359cf3946f