Made checkstyle, PMD, and JUnit ignore files in tinker package, which exists only to try things out, not for production.
Added Owner element/attribute to Source and SourceIndex XML elements in Source XML schema, and rebuilt JAXB classes. Now one can follow ownership of a SensorData back to a Source back to a User.

First pass at an in-memory database implementation. I created this to allow me to push forward on other aspects of the system without getting bogged down in learning CouchDB. Obviously a database that only stores in memory and not disk can't be used in production, but it could be useful for performance testing against CouchDB, and writing a DbImplementation where the persistence was easy allowed me to understand the abstraction layer better, and the requirements. In the course of writing it, ended up making several changes to DbImplementation API. Still needs unit tests.

git-svn-id: https://wattdepot.googlecode.com/svn/trunk@67 962b3a1e-5fa1-11de-82f7-67359cf3946f