Implemented SensorDataResource, but without unit tests right now. This involved many changes to WattDepotResource and other related classes to support the resource. The routing for the SensorDataResource was added to the Server class.
In an attempt to have a very basic demo ready in the near future, added a kludge that creates an admin user and a default source programmatically. This allows me to defer the implementation of the User and Source resources until after the demo.

Split up one test method in TestUserResource.

git-svn-id: https://wattdepot.googlecode.com/svn/trunk@78 962b3a1e-5fa1-11de-82f7-67359cf3946f