Now using Apache commons-cli library and opencsv libraries. Luckily, they were both already in Ivy RoundUp, so grabbed from there. Also added to Eclipse configuration.
Fixed a bug in makeSensorDataRef, the Href element wasn't getting the source name in the URI.

Implemented a tabular data input client. Currently it reads files in the format produced by BuildingManagerOnline for Veris power-only meters, but will be refactored into an abstract class and concrete subclasses for each type of meter datafile.

git-svn-id: https://wattdepot.googlecode.com/svn/trunk@90 962b3a1e-5fa1-11de-82f7-67359cf3946f