First release of the BMODataInputClient. Currently it runs once, sending any stored BMO data (from TSV file) to WattDepot server, and then checking for any new data since the end of the BMO data. If there is no data file, it expects a start time on the command line. Only works for one meter, and should really have a loop option where it will sleep and just check for new data periodically.
To avoid having to put passwords on the command line all the time, created a DataInputClientProperties class to store client properties. Really designed for BMO input, but could be used for other clients.

Since the difference between tabular meter data is the conversion between a row of data into a SensorData object, refactored that into the interface RowParser, and refactored out the one I am using (for the Veris meters used per-floor in Saunders) to VerisRowParser. This should make it easy to write new parsers. In the future, the class name of the parser to be used should probably be specified in a property so that no recompilation would be necessary to have one client for multiple meter types.

Having written two input clients now, there are obvious areas that should be refactored into a parent class. Right now BMO client actually creates a Tabular client and uses it, ick.

GVisualizationServlet now munges time zones such that the DateTimes in the DataTable look like local time in the time zone where the sensor data was created. So if you input HST meter data, when you graph it you see HST times.

In an attempt to track down the cause of a strange problem (Issue 1), added another check on the entity body on PUT of SensorData, and added a test for it. Unfortunately, that didn't solve the problem.

git-svn-id: https://wattdepot.googlecode.com/svn/trunk@97 962b3a1e-5fa1-11de-82f7-67359cf3946f