Wrote Ted5000Sensor that polls a TED 5000 and puts the data into a WattDepot server. Fixes issue 47.
Added support for multiple server processes running on the same system by adding a command line argument to the server indicating the name of the server's home directory under "~/.wattdepot".

Added a convenience method for creating a Property from a double.

Created new jar for TED 5000 sensor.

Bumped version to 1.3.

git-svn-id: https://wattdepot.googlecode.com/svn/trunk@274 962b3a1e-5fa1-11de-82f7-67359cf3946f