Added full support for creating snapshots of the database on the server. This includes changes to the DbImplementation layer, a new resource, a new method in the client library, unit tests, and a client program for initiating the snapshot.
New client (and associated jar file) for making snapshots on the server, intended to be called periodically from a launchd or cron task.

Created new unit test for Derby snapshot code, in special class. Discovered that unit test failures in MemoryStorageImplementation will break the build, but not appear in JUnit reports (due to way db unit tests are run in junit.build.xml). Added comment to that effect.

git-svn-id: https://wattdepot.googlecode.com/svn/trunk@315 962b3a1e-5fa1-11de-82f7-67359cf3946f