Removed race condition in TestDerbyStorageImplementation (hopefully)
TestDerbyStorageImplementation would fail occasionally, but without nay
clear pattern. The snapshot is made asynchronously, so it is possible
that the JUnit assert happened before the snapshot had been made. This
change now tests the backup log file (which should be written to
faster), and also puts a 2 second sleep as a kludge to reduce the
chance of a race condition. Hopefully this fixes things.