Major overhaul of multithreaded sensor code
A variety of issues were improved with the multithreaded sensor:
 - Error handling much improved. The goal is to catch fatal errors that
affect all sensors at top level in MultiThreadedSensor code, and then catch
fatal errors specific to a sensor within the sensor. Connectivity errors
used to be fatal, now they are non-fatal so the sensor should be less
brittle. Fixes issue 105.
 - Much refactoring of common code into MultiThreadedSensor class
 - Refactored initialization code out of run() method
 - Removed redundant initialization code