Support optional Postgres schemas for namespace separation
For Heroku deployment, it is advantagous for applications to be able to
share a single paid database instance. However, the instance only allows
one database catalog and user to be created, which could lead to
namespace collisions between apps. Using Postgres schemas, WattDepot
can be configured to put all tables under a schema, avoiding any
possible namespace collisions. This required some refactoring, because
the schema must be created before any connections are made, and the
initSQL parameter of the Tomcat JDBC pool is used to set the schema
search path for each pooled connection once the schema has been
created. More permanent updates to the search path are possible, but
must be set on a per-user basis, and in the Heroku scenario we only
have one user, so it would defeat the purpose of using schemas in the
first place. Fixes issue #111.