Handle database and schema creation correctly for all cases
The database and schema creation code made several assumptions about
the abilities of the configured Postgres username. These worked fine
if one connected as a user that create databases (like the superuser),
but failed when the user cannot create databases, and is assigned a
single database catalog to work in. This required significant reworking
of the database initialization code, but now should work for superusers
and users without database creation privs both locally and on Heroku.
Fixes issue 121.

Also now supports configuration of the max and initial connection pool
size, which is important on Heroku where the connection pool size for
dev databases is 20. The default initial size is 10, and max is 19 (to
allow for a psql connection to debug things).

Discovered that PreparedStatement is not intended to be used with DDL
commands (like CREATE DATABASE), so replaced with just Statement, after
checking that identifiers are safe.